LayoutTests/platform/ios-sim/storage/indexeddb/intversion-open-in-upgradeneeded-expected.txt

CONSOLE MESSAGE: line 14: TypeError: undefined is not an object (evaluating 'request.onsuccess = deleteSuccess')
Test calling db.open in upgradeneeded.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-open-in-upgradeneeded.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onsuccess = deleteSuccess')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
